start

+++ tokens.hpp
@@ -1,6 +1,5 @@
 #ifndef TOKENS_HPP_
 #define TOKENS_HPP_
-#include <cstdlib>
   enum tokentype
   {
     VOID = 1,
@@ -35,6 +34,6 @@
   };
   extern int yylineno;
   extern char* yytext;
-  extern size_t yyleng;
+  extern int yyleng;
   extern int yylex();
 #endif /* TOKENS_HPP_ */
